requirements.txt
beautifulsoup4==4.12.2
mammoth==1.5.1
openpyxl==3.1.2
pypdfium2==4.7.0
python_pptx==0.6.21
six==1.16.0
xlsx2html==0.4.2
